refactor(viewer): extract shared MiniLineageCard from LibraryRowFlipPopover (VIS-780 / C-4)
No MiniLineageCard existed on-branch; extracted the card body so Library flip, canvas flip, and view-mode flip share one source of truth. Popover reduced to a thin anchoring wrapper.
